Removing save method

+++ __init__.py
@@ -4,10 +4,14 @@ import peewee
 from flask.cli import with_appcontext
 
 db = peewee.SqliteDatabase('people.db')
-
+@click.command("init-db")
+@with_appcontext
+def init_db_command():
+    database = peewee.SqliteDatabase('people.db')
+    database.create_tables([ShoppingRow])
+    print("Initialized the database.")
 
 class ShoppingRow(peewee.Model):
-    id = peewee.AutoField(primary_key=True)
     name = peewee.CharField()
     price = peewee.FloatField()
     quantity = peewee.IntegerField()
@@ -17,7 +21,7 @@ class ShoppingRow(peewee.Model):
 
 
 app = Flask(__name__)
-
+app.cli.add_command(init_db_command)
 
 @app.route("/")
 def all_product():
@@ -27,9 +31,8 @@ def all_product():
 @app.route("/new", methods=["POST"])
 def new_product():
     if "name" in request.form and "price" in request.form and "quantity" in request.form:
-        shopping_row = ShoppingRow.create(name=request.form['name'], price=int(request.form['price']),
+        ShoppingRow.create(name=request.form['name'], price=int(request.form['price']),
                                           quantity=int(request.form["quantity"]))
-        shopping_row.save()
         return render_template("index.html", shopping_list=ShoppingRow.select())
     else:
         return "Il manque des données dans le json", 400
@@ -37,17 +40,14 @@ def new_product():
 
 @app.route("/delete/<int:identifier>", methods=["GET"])
 def get_by_id(identifier: int):
-    shopping_row = ShoppingRow.get(ShoppingRow.id == identifier)
+    shopping_row = ShoppingRow.get_or_none(ShoppingRow.id == identifier)
+    if not shopping_row:
+        return "Il n'existe pas"
     shopping_row.delete_instance()
     return render_template("index.html", shopping_list=ShoppingRow.select())
 
 
-@click.command("init-db")
-@with_appcontext
-def init_db_command():
-    database = peewee.SqliteDatabase('people.db')
-    database.create_tables([ShoppingRow])
-    print("Initialized the database.")
 
 
-app.cli.add_command(init_db_command)
+
+
